requirements.txt
mlx-lm
airllm
torch